requirements.txt
ruamel.yaml>=0.17
pydantic>=2.0
python-dateutil>=2.8
pytest
pytest-cov
black
ruff
mypy
pre-commit
